Removing QuestionService interface.
